requirements.txt
streamlit==1.41.1
langchain==0.1.6
langchain-community==0.0.19
langchain-openai==0.0.5
faiss-cpu==1.7.4
python-dotenv==1.0.0
tiktoken==0.5.2
langsmith==0.0.87
langchain-core==0.1.23
